create four different packages: application, controller, domain, and repository.
The controller will allow to control the API of GET and POST for the ask use case. The domain will store all of the variables and allow for the API to pull the variables that were stored in the class. The repository will be a set of interfaces that will implement SQL with java in order to find the POST of variables in a specific time range, as well as, allow to get all of the POST of ask.