create thankscontroller, userthankscontroller, reportscontroller,notecontroller  classes in controller package with unit tests for them.
The user will be able to create a thanks, update a thanks, and view a thanks.  The thankscontroller will allow to get a list of all the thanks and search for the thanks.

In addition, create note controller and reports controller classes:
NoteController - Using the repo interfaces , the notes controller will allow the user to create notes, updates notes, delete notes, and search notes.
ReportsController - This class will allow the user to get the reports of the