Create conversation, note, noteresult, and thanks classes in the model package.
conversation - this class will contain a string of the user id and a list of all of the notes created by the user.

note - class that will contain the note information that was created by a user.

notesresult - class will contain the conversation list of the user.

thanks - class will contain the thanks information of the user, including description and date created.